+++ tests/test_checker/test_noqa.py
@@ -47,6 +47,7 @@ VERSION_SPECIFIC = types.MappingProxyType({
 
     'WPS416': int(not PY38),  # only works for `< python3.8`
     'WPS451': int(PY38),  # only works for `>= python3.8`
+    'WPS452': int(PY38),  # only works for `>= python3.8`
 
     'WPS602': 2,
 })
@@ -217,7 +218,8 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS449': 1,
     'WPS450': 1,
     'WPS451': 0,  # defined in version specific table.
-    'WPS452': 2,
+    'WPS452': 1,  # also defined in version specific table.
+    'WPS453': 2,
 
     'WPS500': 1,
     'WPS501': 1,